Refactor web module
